// src/types/product.ts
export interface Product {
    id: string;
    brand: string;
    name: string;
    price: number;
    image: string;
    isAvailable: boolean;
    gender: string;
    bestTime: string;
    intensity: string;
    olfactoryFamily: string[];
    notes: string;
    description: string;
    collection: string;
}